+++ tailspin-core/src/main/java/tailspin/language/factory/NodeFactory.java
@@ -24,6 +24,7 @@ import tailspin.language.nodes.array.ArrayMutateNode;
 import tailspin.language.nodes.array.ArrayRangeReadNode;
 import tailspin.language.nodes.array.ArrayReadNode;
 import tailspin.language.nodes.array.PrefixIndexNode;
+import tailspin.language.nodes.array.SuffixIndexNode;
 import tailspin.language.nodes.iterate.ChainNode;
 import tailspin.language.nodes.iterate.CreateRangeNode;
 import tailspin.language.nodes.iterate.ResultAggregatingNode;
@@ -828,12 +829,17 @@ public class NodeFactory {
           sourceCode.createSection(start, end - start));
       case ParseNode(String name, List<?> contents, int start, int end) when name.equals("type-cast") && contents.size() == 2 -> visitTypeCast((ParseNode) contents.getFirst(), (ParseNode) contents.getLast());
       case ParseNode(String name, ParseNode value, int start, int end) when name.equals("prefix-index") -> visitPrefixIndex(value, sourceCode.createSection(start, end - start));
+      case ParseNode(String name, ParseNode value, int start, int end) when name.equals("suffix-index") -> visitSuffixIndex(value, sourceCode.createSection(start, end - start));
       default -> throw new IllegalStateException("Unexpected value: " + source);
     };
   }
 
   private ValueNode visitPrefixIndex(ParseNode value, SourceSection section) {
     return PrefixIndexNode.create(section, asSingleValueNode(visitSource((ParseNode) value.content())));
+  }
+
+  private ValueNode visitSuffixIndex(ParseNode value, SourceSection section) {
+    return SuffixIndexNode.create(section, asSingleValueNode(visitSource((ParseNode) value.content())));
   }
 
   private ValueNode visitTypeCast(ParseNode tag, ParseNode value) {